requirements.txt
fastapi
uvicorn
python-multipart
groq
langchain
sentence-transformers
faiss-cpu
pdfminer.six
tf-keras
langchain_community
langchain_groq
langchain_ollama
bs4
scholarly
google-generativeai
langchain-google-genai
langchain-deepseek
python-dotenv